Final Submission
I know it's not perfect, but I worked really hard and I hope you can appreciate it for what it's worth. This assignment really challenged me in ways I haven't been before so I gained a lot from it.

Thank you again for your understanding during this difficult time, it means a lot.